[ntuple] Add `RNTupleProcessorValuePtr`
When an entry contains missing data for certain fields, for example
because they are missing in a subsequent chain or auxiliary ntuple,
we need a way to signal to the user that the resulting entry has missing
data and is therefore invalid. The `RNTupleProcessorValuePtr`, together
with the `RNTupleProcessorEntry` (which is a thin wrapper around
`REntry`) capture this requirement by only returning actual data if the
entry is marked as "valid" by the processor.