[ntuple] Enforce bare models in RNTupleProcessor
With the addition of `RNTupleValuePtr`, pointers to the values of an
RNTupleModel's default entry, returned by e.g., `MakeField` should not
be used together with the RNTupleProcessor anymore, because they may
contain incorrect data. More specifically, when a particular field's
data could not be loaded for a certain entry, it will retain its
previous value. The `RNTupleValuePtr` adds appropriate checks for this,
but to prevent users from using pointers returned from a model's default
entry, we only allow RNTupleProcessors to be created from bare models.